refactor: finish LandingPage decomposition (F5)
Extract the remaining self-contained sections out of LandingPage into
components/landing/: StepTerminals.js (the step terminals + TerminalTyper
typewriter + STEPS data + BINARY_CHARS) and DashboardPreview.js
(DashboardPreview + MiniChart). LandingPage drops from 944 to 374 lines
and is now a thin orchestrator importing MarqueeTicker, StepTerminals,
and DashboardPreview. Pure code movement; drop the now-unused useRef and
Search imports. 82 tests pass, lint + build clean.
